% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{assert_dataframe_with_data}
\alias{assert_dataframe_with_data}
\title{Assert Object is a Dataframe with at least one row of data}
\usage{
assert_dataframe_with_data(.data)
}
\arguments{
\item{.data}{A Dataframe.}
}
\value{
If successful, returns .data invisibly otherwise throws an error.
}
\description{
Assert Object is a Dataframe with at least one row of data
}
\examples{
assert_dataframe_with_data(mtcars)
}
