% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{is_dataframe_with_data}
\alias{is_dataframe_with_data}
\title{Test if Object is a Dataframe with at least one row of data}
\usage{
is_dataframe_with_data(.data)
}
\arguments{
\item{.data}{A Dataframe.}
}
\value{
a Bool, \code{TRUE} if object is a Dataframe and \code{FALSE}if the
object is not a Dataframe.
}
\description{
Test if Object is a Dataframe with at least one row of data
}
\examples{
is_dataframe_with_data(mtcars)
}
