% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{is_linux}
\alias{is_linux}
\title{Is this a Linux Machine?}
\usage{
is_linux()
}
\value{
A bool; \code{TRUE} if on a linux machine and \code{FALSE}
if not on a linux machine.
}
\description{
Is this a Linux Machine?
}
\examples{
is_linux()
}
