% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{is_windows}
\alias{is_windows}
\title{Is this a Windows Machine?}
\usage{
is_windows()
}
\value{
A bool; \code{TRUE} if on a windows machine and \code{FALSE}
if not on a windows machine.
}
\description{
Is this a Windows Machine?
}
\examples{
is_windows()
}
