% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{make_datekey}
\alias{make_datekey}
\title{Convert a date object into a datekey integer}
\usage{
make_datekey(date)
}
\arguments{
\item{date}{date object}
}
\value{
An integer datekey.
}
\description{
Datekeys can be used to quickly join dates to data
}
\examples{
make_datekey(date = lubridate::make_date())
}
