% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_all_ext}
\alias{read_all_ext}
\title{Read all files with specified extension in a folder}
\usage{
read_all_ext(path, ext)
}
\arguments{
\item{path}{A folder path.}

\item{ext}{A file extension.}
}
\value{
A list of tibbles.
}
\description{
Read all files with specified extension in a folder
}
\examples{
\dontrun{
  read_all_ext("~/data/", "csv")
}
}
\seealso{
Other io: 
\code{\link{read_folder}()}
}
\concept{io}
