% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_common_file}
\alias{read_common_file}
\title{Read a file based on its extension}
\usage{
read_common_file(path, ext)
}
\arguments{
\item{path}{A file path.}

\item{ext}{A file extension.}
}
\value{
A tibble.
}
\description{
Read a file based on its extension
}
