% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{rename_with_stringr}
\alias{rename_with_stringr}
\title{Squish and lowercase all variables in a dataframe}
\usage{
rename_with_stringr(.data)
}
\arguments{
\item{.data}{A Dataframe.}
}
\value{
A Dataframe with the modified variable names.
}
\description{
All variable names in a dataframe will be standardized.
The following is applied:
\itemize{
\item set to lowercase.
\item whitespace at the end and beginning removed.
\item whitespace inside of a string set to a single space.
}
}
\examples{
x <- tibble::tibble(
  ChEeSeSteAk = 1:4,
  "   OH     NO  " = 4:7,
  AHHH = 7:10
)

x |>
  rename_with_stringr()
}
