% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{skip_if_linux}
\alias{skip_if_linux}
\title{Skip unit test if on linux}
\usage{
skip_if_linux()
}
\value{
A message to console if skipping test
}
\description{
Skip unit test if on linux
}
\examples{
\dontrun{
skip_if_not_dph()
}
}
