% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{skip_if_not_windows}
\alias{skip_if_not_windows}
\title{Skip unit test if not on windows}
\usage{
skip_if_not_windows()
}
\value{
A message to console if skipping test
}
\description{
Skip unit test if not on windows
}
\examples{
\dontrun{
skip_if_not_windows()
}
}
