% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{skip_if_windows}
\alias{skip_if_windows}
\title{Skip unit test if not on linux}
\usage{
skip_if_windows()
}
\value{
A message to console if skipping test
}
\description{
Skip unit test if not on linux
}
\examples{
\dontrun{
skip_if_windows()
}

}
