% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/succor-package.R
\docType{package}
\name{succor-package}
\alias{succor}
\alias{succor-package}
\title{succor: Common Solutions for Common Usecases}
\description{
A lightweight set of functions for commonly encountered usecases
}
\seealso{
Useful links:
\itemize{
  \item \url{https://asenetcky-r-pkgs.github.io/succor/}
}

}
\author{
\strong{Maintainer}: Alexander Senetcky \email{alexander.senetcky@ct.gov} (\href{https://orcid.org/0009-0009-3730-5397}{ORCID}) [copyright holder]

}
\keyword{internal}
